Updates to MVTX TB analysis package
The MVTX test beam analysis package has been updated to allow analysis of both test beam data and cosmics data. The cosmic data uses a new decoder, 4 RUs and a different chip map. The user can switch to cosmic analysis by:

        MvtxPrototype2UnpackPRDF *unpack = new MvtxPrototype2UnpackPRDF();
        unpack->isDataTB2019(0);

Default of isDataTB2019 is true so the user doesnt need to change anything to analyse TB data.

NOTE: Make sure that you have the latest decoder, that it is compiled and that it links correctly to the analysis package